requirements.txt
fastapi
uvicorn
requests
python-dotenv
beautifulsoup4
pandas
matplotlib
seaborn
duckdb
lxml
